+++ sources/lusan/CMakeLists.txt
@@ -2,6 +2,7 @@ set(LUSAN "${LUSAN_BASE}/lusan")
 
 include(${LUSAN}/app/CMakeLists.txt)
 include(${LUSAN}/common/CMakeLists.txt)
+include(${LUSAN}/data/CMakeLists.txt)
 include(${LUSAN}/res/CMakeLists.txt)
 include(${LUSAN}/model/CMakeLists.txt)
 include(${LUSAN}/view/CMakeLists.txt)
