+++ sources/lusan/data/CMakeLists.txt
@@ -1,0 +1,4 @@
+include(${LUSAN}/data/common/CMakeLists.txt)
+
+list(APPEND LUSAN_SRC
+)
