+++ sources/lusan/data/common/CMakeLists.txt
@@ -1,0 +1,9 @@
+list(APPEND LUSAN_SRC
+    ${LUSAN}/data/common/OptionsManager.cpp
+    ${LUSAN}/data/common/WorkspaceEntry.cpp
+)
+
+list(APPEND LUSAN_HDR
+    ${LUSAN}/data/common/OptionsManager.hpp
+    ${LUSAN}/data/common/WorkspaceEntry.hpp
+)
