+++ sources/lusan/data/common/WorkspaceEntry.cpp
@@ -10,14 +10,14 @@
  *  with this distribution or contact us at info[at]aregtech.com.
  *
  *  \copyright   © 2023-2024 Aregtech UG. All rights reserved.
- *  \file        lusan/model/common/WorkspaceEntry.cpp
+ *  \file        lusan/data/common/WorkspaceEntry.cpp
  *  \ingroup     Lusan - GUI Tool for AREG SDK
  *  \author      Artak Avetyan
  *  \brief       Lusan application, Dialog to select folder.
  *
  ************************************************************************/
 
-#include "lusan/model/common/WorkspaceEntry.hpp"
+#include "lusan/data/common/WorkspaceEntry.hpp"
 #include <QXmlStreamReader>
 #include <QXmlStreamWriter>
