+++ sources/lusan/model/common/CMakeLists.txt
@@ -1,11 +1,7 @@
 list(APPEND LUSAN_SRC
-    ${LUSAN}/model/common/OptionsManager.cpp
-    ${LUSAN}/model/common/WorkspaceEntry.cpp
     ${LUSAN}/model/common/WorkspaceModel.cpp
 )
 
 list(APPEND LUSAN_HDR
-    ${LUSAN}/model/common/OptionsManager.hpp
-    ${LUSAN}/model/common/WorkspaceEntry.hpp
     ${LUSAN}/model/common/WorkspaceModel.hpp
 )
