+++ sources/lusan/model/common/WorkspaceModel.hpp
@@ -21,7 +21,7 @@
 
 #include <QAbstractListModel>
 #include <vector>
-#include "lusan/model/common/WorkspaceEntry.hpp"
+#include "lusan/data/common/WorkspaceEntry.hpp"
 
 class OptionsManager;
